Add new linter
This linter checks that for function calls there is no space before the parentheses. This is almost the inverse of the  which checks for spaces before parentheses *except* when it is a function.

This only considers function calls and not after , brackets instead of parentheses, or after curly brackets, square brackets, or parentheses. I didn't know if they belonged in this or another linter.

This linter handles cases like  which should be .