Fix crash due to ambiguous cache file paths (#232)
* Renamed "path" to "dir" to avoid ambiguities

Renamed arguments, attributes and variable names in *path to *dir because it was not clear if *path was a file or directory, particularly in cache.R, where the terms "path", "file_path" and "path" were used.

The 'path' argument in functions lint_package() and clear_cache() was deprecated accordingly.

* Added a relative_path() function

...that is more complete and faster than its pre-existing equivalent

* Small doc, code and tests touchups

* relative_path() without "./" prefix

* Fixed ambiguous cache file locations within a package.

Files "R/test.R" and "test.R" were both saved in "cache_dir/test.R". They are now saved in "cache_dir/R/test.R" and "cache_dir/test.R"

* Within the cache folder, use different folders for different packages

This eliminates the possibility that files named the same but in two different packages are cached in the same location.

* Fixed test that failed only under devtools::check()

* Take absolute path for files outside package

* New digest_content() function. Always use mode="list" and inherits=FALSE in get()/exists()

* Fixed bug: duplicate lint when single line content and cache=TRUE

* Restored 'dir' function arguments to their previous 'path' name

and prefer path to dir in variable names

* Shorter solution to the unique cache file problem

Avoid relative paths and their complications. Make a hash of the normalized absolute file name.

* Touchups to make previous commit pass all tests

* Fix bugs in lint_package() exclusions

1/ prevent lint() from crashing by receiving the 'exclusions' parameter twice

2/ make sure exclusions both in the settings and passed as arguments are normalized and relative to the project directory

* More elegant lint_package() exclusion handling

+ some doc upgrades