fixes #252: Allow plain-code-blocks in RMarkdown
Added a test to the end of test.Rmd that included a plain-code-block like this:

```
This shouldn't cause a problem
```

The code-block reproduced the `Malformed File` error described in #252

Plain-blocks match the knitr `chunk.end` pattern at both the start and the end
of the block. So when plain-blocks are present there was 2k _more_ chunk ends
than starts.

A small function `filter_chunk_end_positions` was added to filter the positions
of the chunk-ends based on the chunk-starts