fixes #322: allows Rmd chunks using {lang} engine format
Rmarkdown format allows the use of code blocks in the following format

```{some_engine}
# not necessarily R syntax
foo = bar
```

These blocks should be filtered out before linting the remaining R-code blocks.

A couple of python blocks were added to test.Rmd: one with R-like syntax and one with syntax that can't be parsed as R.

The function `defines_knitr_engine` was extracted from `extract_r_source` and extended. This tests for the presence of {r engine = some_engine} and {some_engine} in the start line of a code chunk. The names of valid engines are obtained from `knitr::knit_engines`