new tests for pipe-continuation linter
pipe-continuation linter was throwing lints in nested expressions.
For example, the nested pipeline in this function:

```
f <- function() {
  this() %>% nested() %>% pipeline()
}
```

This happened because lintr was looking for 'an-expression spanning
several lines, that contained multiple pipe characters on one of
the lines'.

In the above, the definition of `f` is a multi-line expression and
the pipeline is flagged because it is part of that multi-line
expression, even though the pipeline is contained within a
single-line subexpression.