fix #366 : extend pipe-continuation linter to handle nesting (#408)
* new tests for pipe-continuation linter

pipe-continuation linter was throwing lints in nested expressions.
For example, the nested pipeline in this function:

```
f <- function() {
  this() %>% nested() %>% pipeline()
}
```

This happened because lintr was looking for 'an-expression spanning
several lines, that contained multiple pipe characters on one of
the lines'.

In the above, the definition of `f` is a multi-line expression and
the pipeline is flagged because it is part of that multi-line
expression, even though the pipeline is contained within a
single-line subexpression.

* move nested-pipeline tests into main unit tests

* extends pipe-continuation linter to handle nesting

Previous version of pipe-continuation linter threw lints on
single-line pipelines that were nested inside an outer
expression (eg, pipelines in the body of a function, or in a
test_that clause)

The previous version searched for any multi-line expression
that contained several pipe-symbols on the same line.

We now search for any pipe-expression, L %>% R, where;
- no smaller expression surrounds `L %>% R` (to remove
  nesting problems)
- L %>% R spans multiple lines; and
- the expression L contains a pipe-symbol

Then we filter these:
- If R starts on the same line as the pipe-symbol, a lint
  is thrown if L ends on that line
- If R starts on a subsequent line, then a lint is thrown
  if a pipe-symbol in the expression L sits on the same
  line as the pipe-symbol under scrutiny

This has necessitated changing one of the existing tests

`foo %>% bar %>% baz %>%\n qux()`

only throws one pipe-continuation lint now (on the
pipe-symbol before `qux`). Because of how the xml tree is
constructed, the expressions in this are

(foo)
(foo %>% bar)
((foo %>% bar) %>% baz)
(((foo %>% bar) %>% baz) %>% qux)

Of these, the first three expressions are all single-line
pipelines. I can't work out how to extract the whole of
a pipeline as a single xml tree, in a way that doesn't
cause problems with nested expressions.

Have modified the expected value for this test and have
added a number of new tests for pipelines nested in other
expressions.

* update NEWS