match the 'linter'-name to the function name
When Lint() or lint() find a lint, the name of the 'linter' responsible
(in the lintr report) should match the name of the linting function.

For `function_left_parentheses_linter` this was not the case and this
mismatch has now been fixed

For all other linting functions, the reported linter name matches the
function name.