report line-numbers for object-usage lints relative to start-of-file
In #432 it was noted that the line-number for object-usage-linter
warnings were often incorrect. They were being reported relative to the
start of the defining function, rather than the start of the file.

A test was added that throws an object-usage-lint on line-2 of the
function but line-5 of the file.

Also,

- any pre-existing knitr-format tests were expanded so that the observed
lints occurred on an explicit `line_number`

- updated NEWS.md