handle 'expr_or_assign_or_help' in `get_source_expressions`
The token `expr_or_assign_or_help` has been introduced in R-devel (to be
R 4.0).  Parsing of code containing equals-assignments was identified as
a bug when running lintr on R-devel. Here the code for handling
equals-assignments in R3.6 (that checks for the token 'equals_assign')
was extended to check for `expr_or_assign_or_help` (which R-devel uses
in place of `equals_assign`).

Also, code used in an lapply within `get_source_expressions` was split
into a separate function (to aid debugging of `get_source_expressions`).