lintr:::print.lint issue
fixed an issue that caused a
Error in rep.int(character, length) : invalid 'times' value
that caused an issue when printing results from the single_quotes_lintr
It was assumed in highlight_string that the range was such that
range[1] < range[2] which is not necessarily the case when
the string is split over a new line. Added a test case for this
too.