fix #472: allow uneval'd code-blocks with named programming language in Rmd
lintr was complaining that files were malformed when containing
unevaluated code-blocks of the following form (cpp is a format specifier
for use by markdown).
~~~~
```cpp
blah
```
~~~~

"```" matches the end-block pattern for Rmd-code blocks; the change
allows the start pattern for unevaluated blocks to be "```" or
"```some_language"

A couple of such blocks were added to the .Rmd used in the knitr-format
test.

`extract_r_source` has been refactored: selection of start / end
positions for each code block is handled by `get_chunk_positions`.