recursively create cache directory
This should fix #60. As discussed in the links from that issue, by
default ESS turns on caching, which causes lintr to try to create a
subdirectory inside of ~/.R. However, if ~/.R doesn't exist in the
first place, dir.create will fail and lintr will not work in emacs. By
adding the recursive = TRUE argument to dir.create, this will now
create all parent directories as necessary to create the desired cache.