clarify mismatched start/stop exclusion
A user seeing this may not have called `lintr` directly (e.g. it's part of a suite of checks), so specifically calling out that it's related to linting will be helpful.

Also wanted to get the pluralization right, and specifically referring to "ranges" is a further clarification.

I think `covr` has a similar sequence, I will file there shortly.