Clarify seq_linter message
Closes #522

I think basically we need to convey why seq_len is being suggested. I opted for a succinct version that works for all of the `bad_funcs` covered here. I think we could do better with a longer message, but I don't know how much you prefer the messages as terse as possible.