Fix #815 - "error: unexpected input" from unicode character token
Fixes fail obtained when a syntax error interrupts the parsing at a
Unicode character and, as a consequence, the parsed content returned
by 'getParseData(source_file)' includes an incomplete UTF-8 multibyte
representation of that character (which is invalid). See the GitHub
issue discussion for details.

The proposed solution merely "coerces" the getParseData's 'text'
column to what it is supposed to be according to its documentation.

An alternative would be to actually fix the table before use:
> content[!content$terminal, "text"] <- ""