Improve line location information for R parse errors
I chose to use a script in `data-raw` to auto-generate the required parse error regexes and store them in .parse_error_regexes in `R/sysdata.rda`.
`get_source_expressions()` then uses a for-loop after the initial (manual) message checks to see if a parse error with line information was matched.

Fixes #892
Fixes #894

 * [x] implement
 * [x] add tests
 * [x] add NEWS bullet